chore: renamed EL generators
