+++ event-log/src/test/scala/io/renku/eventlog/creation/EventPersisterSpec.scala
@@ -32,7 +32,7 @@ import ch.datascience.graph.model.projects
 import ch.datascience.metrics.{LabeledGauge, TestLabeledHistogram}
 import doobie.implicits._
 import eu.timepit.refined.auto._
-import io.renku.eventlog.DbEventLogGenerators._
+import io.renku.eventlog.EventContentGenerators._
 import io.renku.eventlog._
 import io.renku.eventlog.creation.EventPersister.Result._
 import org.scalamock.scalatest.MockFactory
