+++ event-log/src/test/scala/io/renku/eventlog/init/BatchDateAdderSpec.scala
@@ -29,7 +29,7 @@ import ch.datascience.interpreters.TestLogger
 import ch.datascience.interpreters.TestLogger.Level.Info
 import doobie.implicits._
 import io.circe.literal._
-import io.renku.eventlog.DbEventLogGenerators._
+import io.renku.eventlog.EventContentGenerators._
 import io.renku.eventlog.{CreatedDate, Event}
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
