+++ event-log/src/test/scala/io/renku/eventlog/init/EventStatusRenamerImplSpec.scala
@@ -27,7 +27,7 @@ import ch.datascience.interpreters.TestLogger.Level.Info
 import doobie.implicits._
 import eu.timepit.refined.auto._
 import io.circe.literal.JsonStringContext
-import io.renku.eventlog.DbEventLogGenerators._
+import io.renku.eventlog.EventContentGenerators._
 import io.renku.eventlog._
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
