+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/awaitinggeneration/AwaitingGenerationEventFinderSpec.scala
@@ -32,7 +32,7 @@ import ch.datascience.graph.model.events.{BatchDate, CompoundEventId, EventBody,
 import ch.datascience.graph.model.projects.{Id, Path}
 import ch.datascience.metrics.{LabeledGauge, TestLabeledHistogram}
 import eu.timepit.refined.auto._
-import io.renku.eventlog.DbEventLogGenerators._
+import io.renku.eventlog.EventContentGenerators._
 import io.renku.eventlog._
 import io.renku.eventlog.subscriptions.ProjectIds
 import io.renku.eventlog.subscriptions.awaitinggeneration.ProjectPrioritisation.Priority.MaxPriority
