+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/triplesgenerated/ProjectPrioritisationSpec.scala
@@ -24,7 +24,7 @@ import ch.datascience.graph.model.GraphModelGenerators._
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
 import eu.timepit.refined.numeric.NonNegative
-import io.renku.eventlog.DbEventLogGenerators._
+import io.renku.eventlog.EventContentGenerators._
 import io.renku.eventlog.EventDate
 import io.renku.eventlog.subscriptions.ProjectIds
 import org.scalacheck.Gen
