+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/triplesgenerated/triplesgenerated.scala
@@ -18,9 +18,9 @@
 
 package io.renku.eventlog.subscriptions
 
-import ch.datascience.graph.model.EventsGenerators.compoundEventIds
-import ch.datascience.graph.model.GraphModelGenerators.{projectPaths, projectSchemaVersions}
-import io.renku.eventlog.Generators._
+import ch.datascience.graph.model.EventsGenerators._
+import ch.datascience.graph.model.GraphModelGenerators.projectPaths
+import io.renku.eventlog.EventContentGenerators._
 import org.scalacheck.Gen
 
 package object triplesgenerated {
